Simplify managing investment votes
If I'm right, the `investment_votes` instace variable only exists to
avoid several database queries to get whether the current user has
supported each of the investment.

However, that doesn't make much sense when only one investment is shown.
In this case, the number of queries stays the same, and so we can
simplify the code by rendering the component with an optional parameter.