Fix investment list layout on small/medium screens
We were using a flexible layout, which worked great on large screens.
However, the same cannot be said about small and medium screens; on
those ones, the elements got so small the image isn't even visible and
there was barely any space for the text and so each word took its own
line.

We could do a quickfix and apply the flex layout only on medium screens
and larger; this way we'd get a similar behavior to what we had before
commit 1a73b20d9. However, since now we're using a flexible layout, we
can apply the idea behind Heydon Pickering's flexbox holy albatross [1]
and improve the way these elements are rendered on medium screens.

The idea of this layout is similar to that of container queries
(currently not implemented by any browser [2]), that is, to change the
way the elements are rendered depending on the width of their container,
and not the width of the window. This is particularly handy because
these investments are rendered in the management section as well.

So, assuming a font size of 16px, when the width of the `.panel` element
is greater than the width of the votes (which we manually set to 225px)
plus the width of the investment content (which we set to 250px) plus
the width of the image (140px, defined in the Image model), then all
elements will be on the same line.

And when the width of the `.panel` element is smaller than that, the
`.budget-investment-content` element will take a whole line. Since we've
only got three elements and the content is the one in the middle, that
means each element will be on their own line.

In order to make sure all elements take the whole line in this case, we
set `flex-grow: 1`. And in order to make sure the content element takes
all the available space when all the elements are on the same line, we
set `flex-grow: 1000` for this element.

Currently the code setting the paddings is a bit of a mess because we no
longer use the same styles as in proposals and debates. We'll have to
update those ones too at some point.

Note we can't use the `flex-with-gap` mixin because the container has a
border, and with the negative margin the border would move to the left.

[1] https://heydonworks.com/article/the-flexbox-holy-albatross-reincarnated
[2] https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Container_Queries