Use `RAILS_RELATIVE_URL_ROOT` in place of `CONSUL_RELATIVE_URL`
With this change we can remove part of the code added to keep things
working, like the `kaminari_path` helper, the environments
`relative_url_root` configuration and the GraphQL routes patch.

We can run the application with the relative path as follows:
`RAILS_RELATIVE_URL_ROOT=/aplicaciones/participaciongeneral rails s`

Note that now we've to run assets precompile with the environment
variable `RAILS_RELATIVE_URL_ROOT` to the same value we are going to
use when running the application server so resolutions during compilation
task use the same relative url, otherwise fonts referenced in the fonts.scss
wont be found.

I got inspiration from issue 2435 in the Rails repository.