Remove unneeded lines removing permissions
In the past, users had permission to edit their own legislation
proposals. However, that changed in commit ebfa3fb011, where we replaced
the `can` method with `cannot`.

An easier way to do so is to simplify remove the whole statement, since
by default users don't have permissions to edit their own legislation
proposals.

I'm adding a test for this case, since it was missing.